requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==1.1.0
jinja2==3.1.6
langchain==0.3.26
langchain-core==0.3.83
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain-groq==0.3.8
pinecone==7.3.0
openai==1.109.1
tiktoken==0.12.0
pypdf==5.6.1
sentence-transformers==4.1.0
requests==2.33.0